History screen refresh; Show entries by week and weekday

+++ src/components/History.js
@@ -1,57 +1,15 @@
 import React, { Component } from 'react';
 import update from 'immutability-helper';
-import Infinite from 'react-infinite';
+import classnames from 'classnames';
 import firebase from 'firebase/app';
 import _findIndex from 'lodash/fp/findIndex';
 import _orderBy from 'lodash/fp/orderBy';
+import reverse from 'lodash/reverse';
 import dateFormat from 'date-fns/format';
 import Loader from './Loader';
 
-function EntryListItem({
-	id,
-	timestamp,
-	weight,
-	waist,
-	chest,
-	hips,
-	bf,
-	remove
-}) {
-	timestamp = new Date(timestamp.seconds * 1000);
-	return (
-		<div className="box">
-			<div className="level is-mobile">
-				<div className="level-left">
-					<div className="level-item">
-						<div>
-							<h1 className="title is-5 is-spaced">
-								{timestamp.toDateString()}
-							</h1>
-							<div>{dateFormat(timestamp, 'h:mm a')}</div>
-							{weight && <div>Weight: {weight}</div>}
-							{waist && <div>Waist: {waist}</div>}
-							{chest && <div>Chest: {chest}</div>}
-							{hips && <div>Hips: {hips}</div>}
-							{bf && <div>Bodyfat: {bf}</div>}
-						</div>
-					</div>
-				</div>
-				<div className="level-right">
-					<div className="level-item">
-						<button
-							className="button is-danger is-outlined"
-							onClick={() => {
-								remove(id);
-							}}
-						>
-							Remove
-						</button>
-					</div>
-				</div>
-			</div>
-		</div>
-	);
-}
+import differenceInCalendarWeeks from 'date-fns/differenceInCalendarWeeks';
+import getDay from 'date-fns/getDay';
 
 function UndoDeleteAlert({ undoDelete }) {
 	return (
@@ -62,7 +20,7 @@ function UndoDeleteAlert({ undoDelete }) {
 				bottom: 0,
 				left: 0,
 				width: '100%',
-				zIndex: 1
+				zIndex: 1,
 			}}
 		>
 			<div className="level is-mobile">
@@ -83,12 +41,67 @@ function UndoDeleteAlert({ undoDelete }) {
 	);
 }
 
+function Stat({ value, label, ...props }) {
+	return value ? (
+		<div {...props}>
+			<div className="has-text-grey-light is-uppercase has-text-8">{label}</div>
+			<div>{value}</div>
+		</div>
+	) : null;
+}
+
+function Modal({ entry, closeModal, removeEntry }) {
+	return (
+		<div className="modal is-active px-2">
+			<div className="modal-background" onClick={closeModal}></div>
+			<div className="modal-card">
+				<header className="modal-card-head">
+					<p className="modal-card-title">
+						{dateFormat(entry.timestamp.toDate(), 'MMM d Y h:mm a')}
+					</p>
+					<button
+						className="delete"
+						aria-label="close"
+						onClick={closeModal}
+					></button>
+				</header>
+				<section className="modal-card-body is-flex">
+					<Stat
+						label="Weight"
+						value={entry.weight}
+						className="is-flex-grow-1"
+					/>
+					<Stat label="Waist" value={entry.waist} className="is-flex-grow-1" />
+					<Stat label="Chest" value={entry.chest} className="is-flex-grow-1" />
+					<Stat label="Hips" value={entry.hips} className="is-flex-grow-1" />
+					<Stat label="Bodyfat" value={entry.bf} className="is-flex-grow-1" />
+				</section>
+				<footer className="modal-card-foot">
+					<button
+						className="button is-danger"
+						onClick={() => {
+							closeModal();
+							removeEntry();
+						}}
+					>
+						Remove entry
+					</button>
+					<button className="button" onClick={closeModal}>
+						Cancel
+					</button>
+				</footer>
+			</div>
+		</div>
+	);
+}
+
 let deletedEntry, timeoutHandle;
 
 export default class History extends Component {
 	state = {
 		entries: null,
-		undoDeleteAlert: false
+		undoDeleteAlert: false,
+		showModal: null,
 	};
 	async componentDidMount() {
 		await new Promise((resolve, reject) => {
@@ -101,33 +114,109 @@ export default class History extends Component {
 		this.loadEntries();
 	}
 	render() {
-		const { entries, undoDeleteAlert } = this.state;
-		const entrylistItems = entries
-			? entries.map(e => (
-					<EntryListItem remove={this.removeEntry} key={e.id} {...e} />
-			  ))
+		const { entries, undoDeleteAlert, showModal } = this.state;
+		const entryTable = entries
+			? entries.reduce((result, entry) => {
+					const timestampDate = entry.timestamp.toDate();
+					const week = differenceInCalendarWeeks(
+						timestampDate,
+						entries[entries.length - 1].timestamp.toDate()
+					);
+					const day = getDay(timestampDate);
+
+					if (!result[week]) {
+						result[week] = Array.from({ length: 7 }).map((_) => []);
+					}
+
+					result[week][day] = [...result[week][day], entry];
+					return result;
+			  }, [])
 			: [];
+
 		return (
-			<section className="section">
-				<div className="container">
-					<h1 className="title">
-						History {entries && <span className="tag">{entries.length}</span>}
-					</h1>
-					{entries ? (
-						<Infinite useWindowAsScrollContainer={true} elementHeight={136}>
-							{entrylistItems}
-						</Infinite>
-					) : (
-						<Loader />
-					)}
-					{entries && entries.length < 1 && (
-						<div className="box">No entries yet. Add one to get started.</div>
-					)}
-					{undoDeleteAlert && <UndoDeleteAlert undoDelete={this.undo} />}
-				</div>
-			</section>
+			<div className="container px-2">
+				<h1 className="title">
+					History {entries && <span className="tag">{entries.length}</span>}
+				</h1>
+				{entries ? (
+					<table className="table is-narrow is-bordered is-size-7">
+						<thead>
+							<tr>
+								<th>Sun</th>
+								<th>Mon</th>
+								<th>Tues</th>
+								<th>Wed</th>
+								<th>Thu</th>
+								<th>Fri</th>
+								<th>Sat</th>
+							</tr>
+						</thead>
+						<tbody>
+							{reverse(entryTable).map((week, i) => (
+								<tr key={i}>
+									{week.map((day, i) => (
+										<td
+											key={i}
+											className={classnames('has-text-centered', {
+												'has-background-info-light': !day,
+											})}
+										>
+											{day
+												? day.map(
+														({
+															timestamp,
+															weight,
+															waist,
+															chest,
+															hips,
+															bf,
+															id,
+														}) => (
+															<div
+																key={timestamp}
+																onClick={() => this.showModal(id)}
+															>
+																<div>
+																	{dateFormat(
+																		timestamp.toDate(),
+																		'MMM d Y h:mm a'
+																	)}
+																</div>
+																<Stat label="Weight" value={weight} />
+																<Stat label="Waist" value={waist} />
+																<Stat label="Chest" value={chest} />
+																<Stat label="Hips" value={hips} />
+																<Stat label="Bodyfat" value={bf} />
+															</div>
+														)
+												  )
+												: null}
+										</td>
+									))}
+								</tr>
+							))}
+						</tbody>
+					</table>
+				) : (
+					<Loader />
+				)}
+				{entries && entries.length < 1 && (
+					<div className="box">No entries yet. Add one to get started.</div>
+				)}
+				{undoDeleteAlert && <UndoDeleteAlert undoDelete={this.undo} />}
+				{showModal && (
+					<Modal
+						entry={entries.find((e) => e.id === showModal)}
+						closeModal={() => this.showModal(null)}
+						removeEntry={() => this.removeEntry(showModal)}
+					/>
+				)}
+			</div>
 		);
 	}
+	showModal = (id) => {
+		this.setState({ showModal: id });
+	};
 	// Need to add paging, seems like there's a bug in firestore for paging on a timestamp
 	loadEntries = async () => {
 		const { uid } = firebase.auth().currentUser;
@@ -137,32 +226,28 @@ export default class History extends Component {
 			.collection(`users/${uid}/entries`)
 			.orderBy('timestamp', 'desc')
 			.get()
-			.catch(err => console.error(err));
+			.catch((err) => console.error(err));
 
-		const entries = querySnapshot.docs.map(d =>
+		const entries = querySnapshot.docs.map((d) =>
 			Object.assign(d.data(), { id: d.id })
 		);
-		this.lastEntry = querySnapshot.docs[querySnapshot.docs.length - 1];
 		this.setState({ entries });
 	};
-	removeEntry = async entryId => {
+	removeEntry = async (entryId) => {
 		const { entries } = this.state;
-		const index = _findIndex(e => e.id == entryId)(entries);
+		const index = _findIndex((e) => e.id == entryId)(entries);
 		deletedEntry = entries[index];
-		this.setState(prevState =>
+		this.setState((prevState) =>
 			update(prevState, {
 				entries: { $splice: [[index, 1]] },
-				undoDeleteAlert: { $set: true }
+				undoDeleteAlert: { $set: true },
 			})
 		);
 		const { uid } = firebase.auth().currentUser;
 		timeoutHandle = setTimeout(() => {
 			this.setState({ undoDeleteAlert: false });
 		}, 7000);
-		await firebase
-			.firestore()
-			.doc(`users/${uid}/entries/${entryId}`)
-			.delete();
+		await firebase.firestore().doc(`users/${uid}/entries/${entryId}`).delete();
 	};
 	undo = async () => {
 		const { uid } = firebase.auth().currentUser;
@@ -171,12 +256,12 @@ export default class History extends Component {
 			.collection(`users/${uid}/entries`)
 			.add(deletedEntry);
 
-		this.setState(prevState => {
+		this.setState((prevState) => {
 			const statePatch = update(prevState, {
 				entries: {
-					$push: [Object.assign(deletedEntry, { id: newId })]
+					$push: [Object.assign(deletedEntry, { id: newId })],
 				},
-				undoDeleteAlert: { $set: false }
+				undoDeleteAlert: { $set: false },
 			});
 
 			statePatch.entries = _orderBy(['timestamp'])(['desc'])(
